+++ include/AppVersion.h
@@ -1,3 +1,3 @@
 #pragma once
 
-#define GOODTIME_KIDS_VERSION "2.1.0"
+#define GOODTIME_KIDS_VERSION "3.0.0"